Updated Information flow using Promises
