+++ index.html
@@ -105,8 +105,7 @@
     </div>
     <!-- You can also require other files to run in this process -->
     <script src="./assets/howler.min.js"></script>
-    <script src="./const.js"></script>
-    <script src="./renderer.js"></script>
+    <script src="./global.js"></script>
     <script src="./recent.js"></script>
     <script>
     </script>
